Move services into dedicated directory + get new content from setContent
